% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBAT.fields_by_version.R
\name{GBAT.fields_by_version}
\alias{GBAT.fields_by_version}
\title{Crosstab of return fields available for each version of Geosupport.}
\usage{
GBAT.fields_by_version()
}
\value{
A crosstab data frame containing return fields available for each version of DCP's Geosupport geocoding software.
}
\description{
The \code{GBAT.fields_by_version} function provides a crosstab data table containing return fields available for each version of DCP's Geosupport geocoding software. These fields are described in \href{https://nycplanning.github.io/Geosupport-UPG/appendices/appendix13/}{Appendix 13 of the Geosupport System User Programming Guide}.
}
\details{
Geosupport can be downloaded from \href{https://www1.nyc.gov/site/planning/data-maps/open-data.page#geocoding_application}{DCP BYTES of the BIG APPLE}. Excellent documentation is provide in the \href{https://nycplanning.github.io/Geosupport-UPG/}{Geosupport System User Programming Guide}. The geocoder return fields are described in \href{https://nycplanning.github.io/Geosupport-UPG/appendices/appendix13/}{Appendix 13 of the Geosupport System User Programming Guide}.
}
\examples{

#generate crosstab
gc_fld_by_ver <- GBAT.fields_by_version()

#check which versions output address point id and USPS city name
gc_fld_by_ver[gc_fld_by_ver$GC_fldname \%in\% 
    c("FAP.ap_id","F1E.USPS_city_name"),]

#check which versions output joined fields
gc_fld_by_ver[grepl("^JN",gc_fld_by_ver$GC_fldname),]
}
