% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBAT.get_info.R
\name{GBAT.get_info}
\alias{GBAT.get_info}
\title{Get a list containing return fields, version, and release information for a specific release or version of DCP's Geosupport geocoding software.}
\usage{
GBAT.get_info(GBAT_name)
}
\arguments{
\item{GBAT_name}{the release (e.g., "17B") or version (e.g.,"17.2") of DCP's Geosupport geocoding software as string.  Required.}
}
\value{
A list of available return fields, version, and release information for a specific release or version of DCP's Geosupport geocoding software.
}
\description{
The \code{GBAT.get_info} provides a list of available return fields, version, and release information for a specific release or version of DCP's Geosupport geocoding software. These fields are described in \href{https://nycplanning.github.io/Geosupport-UPG/appendices/appendix13/}{Appendix 13 of the Geosupport System User Programming Guide}.
}
\details{
Geosupport can be downloaded from \href{https://www1.nyc.gov/site/planning/data-maps/open-data.page#geocoding_application}{DCP BYTES of the BIG APPLE}. Excellent documentation is provide in the \href{https://nycplanning.github.io/Geosupport-UPG/}{Geosupport System User Programming Guide}.
}
\examples{
#specify either Geosupport software version or data files release
my_GBAT_info <- GBAT.get_info("17.3")

#Geosupport software version
my_GBAT_info[["GBAT.version"]]

#Geosupport data files release
my_GBAT_info[["GBAT.release"]]

#return fields available
my.GBAT_flds <- my_GBAT_info[["GBAT.return_fields"]]

#preview return fields available for intersection addresses
head(my.GBAT_flds[my.GBAT_flds$GC_function \%in\% c("F2","all"),])

#preview return fields available for standard addresses
head(my.GBAT_flds[my.GBAT_flds$GC_function \%in\% c("FAP","F1A","F1E","all"),])

#text string to search for return fields containing BBL (borough-block-lot)
my.GBAT_flds[grepl("BBL",my.GBAT_flds$GC_fldname,ignore.case=TRUE),]
}
