% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBAT.process_BIN.R
\name{GBAT.process_BIN}
\alias{GBAT.process_BIN}
\title{Geocode a data frame containing a column of BIN with a user selected release or version of DCP's Geosupport.}
\usage{
GBAT.process_BIN(in_df, bin_col_name, 
    source_cols, geocode_fields, return_type = "all", 
    GBAT_name)
}
\arguments{
\item{in_df}{a data frame containing a column of BIN.  Required.}

\item{bin_col_name}{the name of the column containing the input BIN as string.  Required.}

\item{source_cols}{vector of column names from the input data frame to be returned with geocoder results.  Required.}

\item{geocode_fields}{vector of field names generated by the geocoder to be returned with geocoder results.  Required.}

\item{return_type}{option to exclude address which failed to geocode from output as string.  Optional.}

\item{GBAT_name}{the release or version of DCP's Geosupport geocoding software as string.  Required.}
}
\value{
A data frame containing the columns specified in \code{source_cols} plus the geocoder return fields specified in \code{geocode_fields}.
}
\description{
The \code{GBAT.process_BIN} function allows the user to select a release or version of DCP's Geosupport to geocode a data frame containing a column of BIN (building identification numbers).  It aggregates duplicates, geocodes, and then merges the geocoded results back to the original data frame.  The \code{GBAT.process_BIN} function utilizes Geosupport function BN.
}
\details{
Geosupport can be downloaded from \href{https://www1.nyc.gov/site/planning/data-maps/open-data.page#geocoding_application}{DCP BYTES of the BIG APPLE}. Excellent documentation is provide in the \href{https://nycplanning.github.io/Geosupport-UPG/}{Geosupport System User Programming Guide}. The geocoder return fields are described in \href{https://nycplanning.github.io/Geosupport-UPG/appendices/appendix13/}{Appendix 13 of the Geosupport System User Programming Guide}.
}
\examples{
#create a data frame of BIN
in_bin <- 1002376
BIN <- as.character((in_bin-10):(in_bin+10))
u_id <- 1:length(BIN)
df = data.frame(u_id, BIN)

#specify columns from input data frame to retain
source_cols <- c('u_id')

#specify geocoder return fields
geocode_fields <- c("FBN.bbl", "FBN.bin", "FBN.RPAD_lucc",
    "FBN.addr_range_1ax1.bin","FBN.addr_range_1ax1.lo_hse_nbr",
    "FBN.addr_range_1ax1.hi_hse_nbr","FBN.addr_range_1ax1.st_name",
    "FBN.addr_range_1ax2.bin","FBN.addr_range_1ax2.lo_hse_nbr",
    "FBN.addr_range_1ax2.hi_hse_nbr","FBN.addr_range_1ax2.st_name",
    "FBN.addr_range_1ax3.bin","FBN.addr_range_1ax3.lo_hse_nbr",
    "FBN.addr_range_1ax3.hi_hse_nbr","FBN.addr_range_1ax3.st_name")

#geocode data frame by BIN
df_gc <- GBAT.process_BIN(in_df=df, bin_col_name="BIN", 
    source_cols=source_cols, geocode_fields=geocode_fields, 
    return_type="all", GBAT_name="20C")
}
