% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBAT.process_freeform_addresses.R
\name{GBAT.process_freeform_addresses}
\alias{GBAT.process_freeform_addresses}
\title{Geocode a data frame of freeform NYC addresses with a user selected release or version of DCP's Geosupport.}
\usage{
GBAT.process_freeform_addresses(in_df, addr_col_name, 
    unit_col_name = NULL, third_col_name, source_cols, 
    geocode_fields, third_col_type = "boro_code", 
    return_type = "all", GBAT_name)
}
\arguments{
\item{in_df}{a data frame containing NYC addresses.  Required.}

\item{addr_col_name}{the name of the input addresses column as string.  Required.}

\item{unit_col_name}{the name of the unit column as string.  Only available for Geosupport version 16D and later. Optional.}

\item{third_col_name}{the name of either the borough code or zip code column as string.  Required.}

\item{source_cols}{vector of column names from the input data frame to be returned with geocoder results.  Required.}

\item{geocode_fields}{vector of field names generated by the geocoder to be returned with geocoder results.  Required.}

\item{third_col_type}{either \code{"boro_code"} or \code{"zip_code"} as string.  Required.}

\item{return_type}{option to exclude address which failed to geocode from output as string.  Optional.}

\item{GBAT_name}{the release or version of DCP's Geosupport geocoding software as string.  Required.}
}
\value{
A data frame containing the columns specified in \code{source_cols} plus the geocoder return fields specified in \code{geocode_fields}.
}
\description{
The \code{GBAT.process_freeform_addresses} function allows the user to select a release or version of DCP's Geosupport to geocode a data frame of freeform NYC addresses where house number and street name are together in one column.  It aggregates duplicates, geocodes, and then merges the geocoded results back to the original data frame.  The \code{GBAT.process_freeform_addresses} function utilizes Geosupport functions 1A, 1E, and AP.
}
\details{
Geosupport can be downloaded from \href{https://www1.nyc.gov/site/planning/data-maps/open-data.page#geocoding_application}{DCP BYTES of the BIG APPLE}. Excellent documentation is provide in the \href{https://nycplanning.github.io/Geosupport-UPG/}{Geosupport System User Programming Guide}. The geocoder return fields are described in \href{https://nycplanning.github.io/Geosupport-UPG/appendices/appendix13/}{Appendix 13 of the Geosupport System User Programming Guide}.
}
\examples{
#create a data frame of addresses
ADDR <- c("448 Lafayette Ave","727 Manhattan Ave","31-00 47th Ave",
    "308 E 78th S","10 Morton St")
ZIP_CODE <- c("11205","11222","11101","10075","10014")
BORO_CODE <- c(3,3,4,1,1)
u_id <- 1:length(ADDR)
df = data.frame(u_id, ADDR, ZIP_CODE, BORO_CODE)

#specify columns from input data frame to retain
source_cols <- c('u_id')

#specify geocoder return fields
geocode_fields <- c('F1E.output.bin','F1E.output.bbl','F1E.longitude',
    'F1E.latitude','JN.ZCTA_10','F1E.output.ret_code','F1E.output.msg')

#geocode data frame by zip code
df_gc1 <- GBAT.process_freeform_addresses(in_df=df, addr_col_name="ADDR", 
    third_col_name='ZIP_CODE', source_cols=source_cols, geocode_fields=geocode_fields, 
    third_col_type="zip_code", return_type="all", GBAT_name="24B")

#geocode data frame by borough code
df_gc2 <- GBAT.process_freeform_addresses(in_df=df, addr_col_name="ADDR", 
    third_col_name='BORO_CODE', source_cols=source_cols, geocode_fields=geocode_fields, 
    third_col_type="boro_code", return_type="all", GBAT_name="24B")
}
