% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USCB.process_freeform_addresses.R
\name{USCB.process_freeform_addresses}
\alias{USCB.process_freeform_addresses}
\title{Geocode a data frame of freeform US addresses with the USCB batch geocoding service.}
\usage{
USCB.process_freeform_addresses(in_df, addr_col_name,  
    city_col_name, state_col_name, zipcode_col_name, 
    rec_num = 10000)
}
\arguments{
\item{in_df}{a data frame or data table containing US addresses.  Required.}

\item{addr_col_name}{the name of the input addresses column as string.  Required.}

\item{city_col_name}{the name of the input city column as string.  Required.}

\item{state_col_name}{the name of the input state column as string.  Required.}

\item{zipcode_col_name}{the name of the input zip code column as string.  Required.}

\item{rec_num}{number of records to process with each hit to the geocoding service. Optional.}
}
\value{
A data frame or data table (depending on input class) containing the original data frame plus the geocoder return fields which include: block, tract, PUMA, TigerLine, coordinates and matching information for 2000, 2010, and current geographic vintages.
}
\description{
The \code{USCB.process_freeform_addresses} function allows the user to geocode a data frames of freeform US addresses where house number and street name are together in one column.  It aggregates duplicate records, geocodes using the US Census Bureau batch geocoding service, and then merges the geocoded results back to the original data frame.
}
\details{
For more information, refer to the \href{https://www.census.gov/geo/maps-data/data/geocoder.html}{Census Geocoder webpage}.
}
\examples{
#create a data frame of addresses
#we'll use the NYC OpenData Restaurant Inspections API
URL <- "https://data.cityofnewyork.us/resource/43nn-pn8j.json"
str1 <- "$where=(inspection_date>='2019-01-01')"
str2 <- "$group=camis,dba,building,street,boro,zipcode"
str3 <- "$select=camis,dba,building,street,boro,zipcode"
str4 <- "$limit=20000"
data_URL <- paste0(URL,"?",str1,"&",str2,"&",str3,"&",str4)
con <- url(URLencode(data_URL))
data_JSON <- jsonlite::fromJSON(con)
data_JSON$ADDR <- paste(data_JSON$building, data_JSON$street, sep=" ")
data_JSON$STATE <- "NY"
 
#test batch process speed
test.df1 <- data_JSON[1:100,]
ptm <- proc.time()
GEO_df2a <- USCB.process_freeform_addresses(test.df1, "ADDR", "boro", 
  "STATE", "zipcode", 10)
proc.time() - ptm

#preview results
head(GEO_df2a)
}
